cloog_domain_from_isl_set: turn parameter sets into regular sets
Recent versions of isl make an explicit distinction between regular
sets and parameter sets, but CLooG does not make any such distinction.
Usually, this doesn't lead to any problems because the context
(which should be a parametric set) is usually mapped into a higher
dimension before being combined with other sets (iteration domains).
This extension automatically turns a parameter set into a set.
However, when the iteration domains are zero-dimensional,
cloog_loop_restrict will not call cloog_domain_extend on the original
context, but will instead directly intersect the iteration domain
with the context.

We could simplify the code of cloog_loop_restrict to always call
cloog_domain_extend, but that may have a negative effect on other
backends.  Instead, we choose to turn any parametric set into
a regular set while it is being converted to a CloogDomain.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>
Signed-off-by: Tobias Grosser <tobias@grosser.es>