Add support for OpenSCop Coordinates extension
This patch allows a complete source-to-source experience
with CLooG when using OpenScop's "Coordinates" extension. When
this extension is provided, the generated code is now injected by
the pretty-printer in the code pointed by the extension at the
place of a given SCoP.

Note that since the osl submodule is also updated with this patch,
it also solves a building problem with GMP configuration which used
to be different in CLooG and osl, as reported by rene.sugar.