Loop execution time reporting
Set clast_for->time_var_name to the name of the variable which will
accumulate execution time of the body inside this loop (in seconds
double precision floating-point) if macro TIME is defined (-DTIME).
The code is linked to cloog_util_rtclock() that relies on gettimeofday
and is part of the CLooG library.

Signed-off-by: Uday Bondhugula <udayreddy@gmail.com>
Signed-off-by: Cedric Bastoul <cedric.bastoul@u-psud.fr>