test/checker.sh: various improvements
The previous implementation would decide that a test fails as long as
Valgrind detects memory still in use at exit. The current implementation
checks whether this is due to actual memory leaks and lets the test pass
if the memory is still reachable.

This commit also changes the output of the script `checker.sh`. The
output is more compact as most debug output is now redirected to log
files (which can be found in the directory `test/logs`).

Colour printing now uses `printf` instead of echo: the `-e` option is
useless (and might not even exist) in POSIX-compliant implementations of
`/bin/sh` (such as Dash, for instance) whereas it is necessary with
other shells (such as Bash). With `printf`, both types of shell (at
least Dash and Bash) correctly interpret the escape sequences used for
colours.

The previous implementation used hardcoded names for temporary files
which could prevent the script to be run multiple times in parallel.
With this commit, the script uses names that should be specific to a
given instance of the script.

This also adds support for openscop input in generate_test_advanced.

Signed-off-by: Harenome Ranaivoarivony Razanajato <razanajato@etu.unistra.fr>
Signed-off-by: Maxime Schmitt <maxime.schmitt91@gmail.com>