CLooG integer representation selection
This patch introduce the comeback of CLooG int selection.

During configuration time you can choose whatever representation you prefer
(defaulting to GMP).

Because some functions of the library use cloog_int as a parameter, it is not
possible to change the precision once the library is compiled. This will break
calling conventions with types of different size.

You can choose the precision by adding "--with-cloog-int=xxx" where xxx is the
string "gmp" or "int" or "long" or "longlong".

Signed-off-by: Maxime Schmitt <maxime.schmitt91@gmail.com>