Adding ISL imath and imath-32 support
Configure can now search for ISL integer representation when ISL is used and
*not bundled*. This allows the use imath or imath-32 instead of GMP without
having CLooG complaining.

We looks for the file "isl/val_gmp.h" that is only present when ISL is compiled
with GMP support.

The only portable way to transfer a value between GMP and IMath without loosing
precision is to print the value as a string. Thus the bundled default still
uses GMP. This ensure that both CLooG and ISL use the same representation
avoiding useless allocations and parsing of strings when a value need to be
passed from CLooG to ISL and vis-versa.

Signed-off-by: Maxime Schmitt <maxime.schmitt91@gmail.com>